requirements.txt
python-pptx
markdown
beautifulsoup4
pydantic
requests
python-dotenv
